+++ ecd-20200930/src/Models/ModifyOfficeSiteDnsInfoRequest.php
@@ -1,0 +1,89 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Ecd\V20200930\Models;
+
+use AlibabaCloud\Dara\Model;
+
+class ModifyOfficeSiteDnsInfoRequest extends Model
+{
+    /**
+     * @var string[]
+     */
+    public $dnsAddress;
+    /**
+     * @var string
+     */
+    public $officeSiteId;
+    /**
+     * @var string
+     */
+    public $regionId;
+    protected $_name = [
+        'dnsAddress'   => 'DnsAddress',
+        'officeSiteId' => 'OfficeSiteId',
+        'regionId'     => 'RegionId',
+    ];
+
+    public function validate()
+    {
+        if (\is_array($this->dnsAddress)) {
+            Model::validateArray($this->dnsAddress);
+        }
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->dnsAddress) {
+            if (\is_array($this->dnsAddress)) {
+                $res['DnsAddress'] = [];
+                $n1                = 0;
+                foreach ($this->dnsAddress as $item1) {
+                    $res['DnsAddress'][$n1++] = $item1;
+                }
+            }
+        }
+
+        if (null !== $this->officeSiteId) {
+            $res['OfficeSiteId'] = $this->officeSiteId;
+        }
+
+        if (null !== $this->regionId) {
+            $res['RegionId'] = $this->regionId;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['DnsAddress'])) {
+            if (!empty($map['DnsAddress'])) {
+                $model->dnsAddress = [];
+                $n1                = 0;
+                foreach ($map['DnsAddress'] as $item1) {
+                    $model->dnsAddress[$n1++] = $item1;
+                }
+            }
+        }
+
+        if (isset($map['OfficeSiteId'])) {
+            $model->officeSiteId = $map['OfficeSiteId'];
+        }
+
+        if (isset($map['RegionId'])) {
+            $model->regionId = $map['RegionId'];
+        }
+
+        return $model;
+    }
+}